requirements.txt
customtkinter
httpx
pyperclip
openai
Pillow
pystray
